+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+## [6.3.2] - 2025-12-16
+
+* Use call_nowait() for service only if it is ready
+
 ## [6.3.1] - 2025-12-15
 
 * Expose ProtocolViolationError info